requirements.txt
streamlit==1.28.0
keras-ocr==0.9.3
numpy==1.26.2
python-dotenv==1.0.0
together
pypdf2==3.0.1
python-multipart==0.0.6
tensorflow